requirements.txt
python-telegram-bot==20.3
Flask==3.0.3
SQLAlchemy==2.0.32
python-dotenv==1.0.1
alembic==1.13.2
pydantic==2.8.2
